requirements.txt
tornado==4.4.2
jinja2==2.8.1
peewee==2.8.5
BeautifulSoup==3.2.1
pycrypto==2.6.1
MySQL-python==1.2.5
